Refactor cart feature and implement CartPage step definitions
