+++ src/test/java/com/amazon/runners/TestRunner.java
@@ -12,7 +12,7 @@ import org.junit.runner.RunWith;
         glue = "com.amazon.stepdefinitions",
         monochrome = true,
         dryRun = false,
-        tags = "@cart"
+        tags = ""
 )
 public class TestRunner {
