+++ src/test/java/com/amazon/utilities/WebDriverUtils.java
@@ -33,4 +33,6 @@ public class WebDriverUtils {
         return new WebDriverWait(driver, Duration.ofSeconds(timeout))
                 .until(ExpectedConditions.visibilityOf(element));
     }
+
+
 }
